+++ core/src/common/index.ts
@@ -2,3 +2,4 @@ export * from "./decorators/index.js";
 export * from "./classes/index.js";
 export * from "./constants/index.js";
 export * from "./typings/index.js";
+export * from "./utils/index.js";