+++ core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nailyjs/core
 
+## 0.30.4
+
+### Patch Changes
+
+- d25d072: feat: create command & change @nailyjs/vue 's scripts
+
 ## 0.30.3
 
 ### Patch Changes