+++ tests/sample/CHANGELOG.md
@@ -1,5 +1,12 @@
 # sample
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [d25d072]
+  - @nailyjs/core@0.30.4
+
 ## 1.0.1
 
 ### Patch Changes