+++ tests/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # example-vue
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [d25d072]
+  - @nailyjs/vue@0.30.4
+  - @nailyjs/core@0.30.4
+
 ## 0.0.1
 
 ### Patch Changes